# lib/api-spec/openapi.yaml
openapi: 3.1.0
info:
  # Do not change the title, if the title changes, the import paths will be broken
  title: Api
  version: 0.1.0
  description: Smart Farmer Assistance API
servers:
  - url: /api
    description: Base API path
...
paths:
...
  /auth/demo-login:
    post:
      operationId: demoLogin
      tags: [auth]
      summary: Sign in to a demo role
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/DemoLoginInput"
      responses:
        "200":
          description: Signed in
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Session"
...
  /profile:
    get:
      operationId: getProfile
      tags: [farmers]
      summary: Get the current farmer profile
      responses:
        "200":
          description: Farmer profile
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/FarmerProfile"
    patch:
      operationId: updateProfile
      tags: [farmers]
      summary: Update the current farmer profile
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/ProfileUpdate"
      responses:
        "200":
          description: Updated profile
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/FarmerProfile"
...
  /centres/{centreId}/slots:
    get:
      operationId: listCentreSlots
      tags: [centres]
      summary: List available booking slots for a centre
      parameters:
        - $ref: "#/components/parameters/CentreId"
      responses:
        "200":
          description: Available slots
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/Slot"
  /bookings:
    get:
      operationId: listBookings
      tags: [bookings]
      summary: List bookings
      parameters:
        - in: query
          name: status
          schema:
            type: string
            enum: [all, upcoming, completed, cancelled, pending]
        - in: query
          name: search
          schema:
            type: string
      responses:
        "200":
          description: Bookings
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/Booking"
    post:
      operationId: createBooking
      tags: [bookings]
      summary: Book a procurement slot
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/BookingInput"
      responses:
        "201":
          description: Booking created
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Booking"
        "409":
          description: Slot unavailable
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Error"
  /bookings/{bookingId}:
    get:
      operationId: getBooking
      tags: [bookings]
      summary: Get booking details
      parameters:
        - $ref: "#/components/parameters/BookingId"
      responses:
        "200":
          description: Booking
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Booking"
    patch:
      operationId: updateBooking
      tags: [bookings]
      summary: Update booking status
      parameters:
        - $ref: "#/components/parameters/BookingId"
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/BookingUpdate"
      responses:
        "200":
          description: Updated booking
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Booking"
    delete:
      operationId: cancelBooking
      tags: [bookings]
      summary: Cancel a booking
      parameters:
        - $ref: "#/components/parameters/BookingId"
      responses:
        "200":
          description: Cancelled booking
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Booking"
...
  /notifications/{notificationId}/read:
    patch:
      operationId: markNotificationRead
      tags: [notifications]
      summary: Mark notification as read
      parameters:
        - in: path
          name: notificationId
          required: true
          schema:
            type: string
      responses:
        "200":
          description: Updated notification
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Notification"
...
components:
  parameters:
    CentreId:
      in: path
      name: centreId
      required: true
      schema:
        type: string
    BookingId:
      in: path
      name: bookingId
      required: true
      schema:
        type: string
  schemas:
...
    Error:
      type: object
      properties:
        error:
          type: string
      required: [error]
    DemoLoginInput:
      type: object
      properties:
        role:
          type: string
          enum: [farmer, officer, admin]
      required: [role]
    Session:
      type: object
      properties:
        user:
          $ref: "#/components/schemas/User"
        token:
          type: string
      required: [user, token]
    User:
      type: object
      properties:
        id:
          type: string
        name:
          type: string
        role:
          type: string
          enum: [farmer, officer, admin]
        location:
          type: string
        initials:
          type: string
      required: [id, name, role, location, initials]
    FarmerProfile:
      type: object
      properties:
        id: { type: string }
        name: { type: string }
        mobile: { type: string }
        village: { type: string }
        district: { type: string }
        state: { type: string }
        landArea: { type: number }
        crops:
          type: array
          items: { type: string }
        expectedYield: { type: string }
        maskedAccount: { type: string }
        completion: { type: integer }
      required: [id, name, mobile, village, district, state, landArea, crops, expectedYield, maskedAccount, completion]
    ProfileUpdate:
      type: object
      properties:
        name: { type: string, minLength: 2 }
        village: { type: string }
        district: { type: string }
        state: { type: string }
        landArea: { type: number, minimum: 0 }
        crops:
          type: array
          items: { type: string }
        expectedYield: { type: string }
...
    ProcurementCentre:
      type: object
      properties:
        id: { type: string }
        name: { type: string }
        address: { type: string }
        district: { type: string }
        distance: { type: string }
        open: { type: boolean }
        queue: { type: integer }
        capacity: { type: integer }
        remaining: { type: integer }
        waitMinutes: { type: integer }
        phone: { type: string }
        coordinates:
          type: object
          properties:
            lat: { type: number }
            lng: { type: number }
          required: [lat, lng]
      required: [id, name, address, district, distance, open, queue, capacity, remaining, waitMinutes, phone, coordinates]
    Slot:
      type: object
      properties:
        id: { type: string }
        date: { type: string }
        label: { type: string }
        available: { type: integer }
        capacity: { type: integer }
        recommended: { type: boolean }
      required: [id, date, label, available, capacity, recommended]
    BookingInput:
      type: object
      properties:
        crop: { type: string, minLength: 2 }
        quantity: { type: number, minimum: 1 }
        centreId: { type: string }
        slotId: { type: string }
      required: [crop, quantity, centreId, slotId]
    BookingUpdate:
      type: object
      properties:
        status:
          type: string
          enum: [pending, confirmed, approaching, completed, cancelled]
    Booking:
      type: object
      properties:
        id: { type: string }
        crop: { type: string }
        quantity: { type: number }
        centre:
          $ref: "#/components/schemas/ProcurementCentre"
        date: { type: string }
        time: { type: string }
        status:
          type: string
          enum: [pending, confirmed, approaching, completed, cancelled]
        arrival: { type: string }
        quality: { type: string }
        procurement: { type: string }
        payment: { type: string }
      required: [id, crop, quantity, centre, date, time, status, arrival, quality, procurement, payment]
...
    Notification:
      type: object
      properties:
        id: { type: string }
        title: { type: string }
        message: { type: string }
        time: { type: string }
        type: { type: string }
        read: { type: boolean }
      required: [id, title, message, time, type, read]
...