# lib/api-zod/src/generated/types/booking.ts
/**
 * Generated by orval v8.23.0 🍺
 * Do not edit manually.
 * Api
 * Smart Farmer Assistance API
 * OpenAPI spec version: 0.1.0
 */
import type { BookingStatus } from './bookingStatus';
import type { ProcurementCentre } from './procurementCentre';

export interface Booking {
  id: string;
  crop: string;
  quantity: number;
  centre: ProcurementCentre;
  date: string;
  time: string;
  status: BookingStatus;
  arrival: string;
  quality: string;
  procurement: string;
  payment: string;
}