# lib/api-zod/src/generated/types/bookingInput.ts
/**
 * Generated by orval v8.23.0 🍺
 * Do not edit manually.
 * Api
 * Smart Farmer Assistance API
 * OpenAPI spec version: 0.1.0
 */

export interface BookingInput {
  /** @minLength 2 */
  crop: string;
  /** @minimum 1 */
  quantity: number;
  centreId: string;
  slotId: string;
}