# lib/api-zod/src/generated/types/demoLoginInput.ts
/**
 * Generated by orval v8.23.0 🍺
 * Do not edit manually.
 * Api
 * Smart Farmer Assistance API
 * OpenAPI spec version: 0.1.0
 */
import type { DemoLoginInputRole } from './demoLoginInputRole';

export interface DemoLoginInput {
  role: DemoLoginInputRole;
}