# lib/api-zod/src/generated/types/farmerProfile.ts
/**
 * Generated by orval v8.23.0 🍺
 * Do not edit manually.
 * Api
 * Smart Farmer Assistance API
 * OpenAPI spec version: 0.1.0
 */

export interface FarmerProfile {
  id: string;
  name: string;
  mobile: string;
  village: string;
  district: string;
  state: string;
  landArea: number;
  crops: string[];
  expectedYield: string;
  maskedAccount: string;
  completion: number;
}