# lib/api-zod/src/generated/types/procurementCentre.ts
/**
 * Generated by orval v8.23.0 🍺
 * Do not edit manually.
 * Api
 * Smart Farmer Assistance API
 * OpenAPI spec version: 0.1.0
 */
import type { ProcurementCentreCoordinates } from './procurementCentreCoordinates';

export interface ProcurementCentre {
  id: string;
  name: string;
  address: string;
  district: string;
  distance: string;
  open: boolean;
  queue: number;
  capacity: number;
  remaining: number;
  waitMinutes: number;
  phone: string;
  coordinates: ProcurementCentreCoordinates;
}