# lib/api-zod/src/generated/types/profileUpdate.ts
/**
 * Generated by orval v8.23.0 🍺
 * Do not edit manually.
 * Api
 * Smart Farmer Assistance API
 * OpenAPI spec version: 0.1.0
 */

export interface ProfileUpdate {
  /** @minLength 2 */
  name?: string;
  village?: string;
  district?: string;
  state?: string;
  /** @minimum 0 */
  landArea?: number;
  crops?: string[];
  expectedYield?: string;
}